pyproject.toml
[tool.poetry]
name = "sign-lanugage-translator"
version = "0.1.0"
description = "Realizacja projektu w ramach przedmiotu Modelowanie i analiza systemów informatycznych"
authors = ["Wojciech Jasiewicz <mail>; Karol Ciepły <mail>; Kacper Moll kacmol418@polsl.pl"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
pydantic = "^2.10.3"
ruff = "^0.8.4"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.7.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.10.1"

[tool.poetry.group.backend.dependencies]
fastapi = "^0.115.6"
uvicorn = "^0.34.0"

[tool.poetry.group.frontend.dependencies]
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version="py312"
exclude=[".venv"]
line-length=120
indent-width=4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
    # Rules list: https://docs.astral.sh/ruff/rules/
    "ALL"
]

ignore = [
    "PLR2004",
    "BLE001",
    "SIM115",
    "S102",
    "RET504",
    "TRY300",
    "PTH123",
    "PD901",
    "W293",
    "PGH003"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

